Require column name(s) in sort-by
